📡 API now triggers surveyform webhook when reinitializing
