+++ api/.env.example
@@ -35,7 +35,11 @@ GITHUB_PATH_ENTITIES=devographics/entities
 # openssl rand -base64 32
 ENCRYPTION_KEY=dk20f9xxxx
 # Secret key that is used to reload the API
+# NOTE: we currently suppose it's the same as surveyform SECRET_KEY
 SECRET_KEY=dev-secret-key-replace-me
+
+# Needed to trigger cache refresh hooks on surveyform
+SURVEYFORM_URL=http://localhost:3000
 
 # Local Dev
