+++ api/src/init.ts
@@ -80,4 +80,18 @@ export const reinitialize = async ({ context, initList = defaultInitList }: Init
     // June 2023: we do not "warm" the cache from API app anymore,
     // it is now the responsibility of each app to handle its own cache
     // await initDbCache({ context, data, initList })
+
+    // However, we still inform the surveyform that a refresh is needed, via it's API
+    // TODO: in the future this could be made unnecessary if surveyform consumed the values set by the API more directly
+    try {
+        if (!process.env.SURVEYFORM_URL) throw new Error("SURVEYFORM_URL not set")
+        // assuming API and surveyform use the same secret key for hooks
+        const resetUrl = `${process.env.SURVEYFORM_URL}/api/cache/refresh-cache?key=${process.env.SECRET_KEY}&${initList.map(item => encodeURIComponent(item)).join("&")}`
+        // NOTE: it's ok to log this secret key as it's already included in the URL,
+        // it's not used to secure any user data just to protect API calls from abuses
+        console.log("Resetting surveyform on URL:", resetUrl)
+        await fetch(resetUrl)
+    } catch (err) {
+        console.warn("Could not reinitialize surveyform cache:", err)
+    }
 }
