+++ surveyform/src/app/api/cache/flush-cache/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-import { flushCache } from "@devographics/fetch";
+import { flushInMemoryCache } from "@devographics/fetch";
 
 /**
  * Empties the short-lived in-memory cache
@@ -9,7 +9,7 @@ export async function GET(req: NextRequest) {
   if (key !== process.env.SECRET_KEY) {
     return NextResponse.json({ error: "invalid_key" });
   } else {
-    flushCache();
+    flushInMemoryCache();
     return NextResponse.json({ data: { result: "ok" } });
   }
 }
